Move test to FacetUtil
